pyproject.toml
[project]
name = "forge"
version = "0.9.0"
description = "Local AI coding assistant with transparent context management"
requires-python = ">=3.10"
license = {text = "Proprietary"}
readme = "README.md"
dependencies = [
    "requests>=2.28",
    "numpy>=1.24",
    "Pillow>=9.0",
    "customtkinter>=5.2",
    "tiktoken>=0.7",
    "pyyaml>=6.0",
    "pyreadline3>=3.4; sys_platform == 'win32'",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0",
    "sounddevice>=0.4",
    "pynput>=1.7",
    "edge-tts>=6.0",
    "soundfile>=0.12",
]

[project.scripts]
forge = "forge.__main__:main"

[tool.pytest.ini_options]
markers = [
    "live: requires a running Ollama instance (use --live to run)",
    "stub_only: requires scripted OllamaStub responses (skipped in --live mode)",
]
timeout = 300

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
